Added an extra, commented out function on ServerProxy. Not Needed for Phase one but will be helpful later.
Completed poller.poll(). It iterates through all the lists of games in the client model and checks for changes. If a change is found, it is added and recorded in a changelist. Same with the user member.

Lots of functions added to ClientModel. Presenters now update ClientModel through ClientFacade.

Presenter Update functions still need to be done.